\name{simSummary-package}
\alias{simSummary-package}
\docType{package}
\title{Simulation summary}
\description{simSummary is a small utility package that eases the process of summarizing simulation results. Simulations often produce intermediate results - some focal statistic(s), that need to be summarized over many replicates. simSummary eases the process of summarizing these focal statistics.}
\details{
\tabular{ll}{Package: \tab simSummary\cr
Title: \tab Simulation summary\cr
Author: \tab Gregor Gorjanc\cr
Maintainer: \tab Gregor Gorjanc <gregor.gorjanc@bf.uni-lj.si>\cr
License: \tab GPL (>= 2)\cr
Version: \tab 0.1.0\cr
Depends: \tab abind (>= 1.4-0), svUnit (>= 0.7-5)\cr
Imports: \tab gdata (>= 2.8.0)\cr
Date: \tab Check NEWS file for changes: news(package='simSummary')\cr}
}
\author{Gregor Gorjanc}

\keyword{ package }


