\name{simSummary_unitTests}
\alias{simSummary_unitTests}
\title{Run unit tests for package simSummary}
\description{Package \pkg{simSummary} has unit tests to test the validity of code behaviour.
Any developeR or useR can easily run these tests as shown in examples bellow. }
\usage{
simSummary_unitTests()
}



\author{Gregor Gorjanc}




\examples{

## Run unit tests for package simSummary
simSummary_unitTests()

}
